Keep transcoders on a list in registration order.
This fixes a bug where it was possible for there to be a transcoder in position
"1" but not in position "0" if a transcoder hardware driver was loaded,
unloaded, and reloaded again without also reloading dahdi_transcode.  The result
is that codec_dahdi fails to enumerate all the transcoders in the system.

(closes issue #14627)
Reported by: xblurone

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@6114 a0bf4364-ded3-4de4-8d8a-66a801d63aff