Do not handle duplicate reponses.
It is possible for system activity to prevent the wctc4xxp driver from
responding the the firmware on the TC400M for a period of time.  If this occurs,
the TC400M will resend a response to a command that we sent.  This duplicate
reponse will then sometimes confuse the driver.  Normally this would manifest
as an attempt to connect the same channel together in a transcoding session.

git-svn-id: https://origsvn.digium.com/svn/dahdi/linux/trunk@6224 22b186d9-ae4d-0410-a2cf-b1b22c4d6fbd