Do not handle duplicate reponses.
It is possible for system activity to prevent the wctc4xxp driver from
responding the the firmware on the TC400M for a period of time.  If this occurs,
the TC400M will resend a response to a command that we sent.  This duplicate
reponse will then sometimes confuse the driver.  Normally this would manifest
as an attempt to connect the same channel together in a transcoding session.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@6224 a0bf4364-ded3-4de4-8d8a-66a801d63aff