wcte12xp: Update cmdqueue processing.
The command queue for reading from the registers on the framer is now stored in
a linked_list instead of an array.  Allows for the locks to protect this
structure to be held for shorter periods of time and reduces the need to cycle
through all the elements in the array to decide if there is a command in the
queue to process. Remove the usecount and dead members from struct t1 since the
module reference count will allow us to know when it's safe to free up the
memory. This change also moves alarm processing out of the interrupt handler and
removes the need for special interrupt handling of commands.

git-svn-id: https://origsvn.digium.com/svn/dahdi/linux/trunk@6525 22b186d9-ae4d-0410-a2cf-b1b22c4d6fbd