voicebus: Create workqueue for each vpmadt032 instance.
Depending on the system latency, the deferred work for the vpmadt032 can take
up to 200ms.  This change allows each vpmadt032 to use its own workqueue, and
not the global system workqueue.  This prevents vpm operations from blocking
the main system workqueue for extended periods.

This restores the behavior to the way it was before the common vpmadt032 code
was moved out of the wctdm24xxp and wcte12xp drivers.

DAHDI-260

voicebus-squash:  Adding the wq name.

git-svn-id: https://origsvn.digium.com/svn/dahdi/linux/trunk@6572 22b186d9-ae4d-0410-a2cf-b1b22c4d6fbd