dahdi-base: Add support for core timing.
This essentially moves the function of dahdi_dummy into the core of DAHDI.  It
ensures that if DAHDI is loaded, it will always be able to provide timing,
regardless of whether there are board drivers loaded, or if the board drivers
are properly calling dahdi_receive.

If there is a master span loaded which is calling dahdi_receive, then the
behavior will be like it is normally.

This functionality is off by default, uncomment CONFIG_DAHDI_CORE_TIMER in
include/dahdi/config_dahdi.h in order to enable it.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@6863 a0bf4364-ded3-4de4-8d8a-66a801d63aff