wctdm24xxp, wctdm: Detect if our hookstate has been set back to the initial state.
Check if our hookstate has been set back to the initial state, typically the
result of a chanconfig, and if so, if we're an FXO port, forget our current
battery state.  This allows the driver to determine and report again what the
hook state of the port is.

(related to issue #14577)
(closes issue #15429)

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7003 a0bf4364-ded3-4de4-8d8a-66a801d63aff