wcfxo: Reset the DAA on module initialization.
The X100p and clones will sometimes work and sometimes not depending on wether
the DAA powers up in running state- this seems to be related to the power
supply. This problem is caused by the driver not reseting the DAA and may be
the source of a great many intermittent problems with this card.

(closes issue #14232)
Reported by: tallen8840
Patch by: tallen8840
Tested by: explidous, Flavio

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7008 a0bf4364-ded3-4de4-8d8a-66a801d63aff