Ensure that dahdi_scan correctly reports VPM presence.
The wct4xxp driver (for the dual/quad span T1/E1 cards) and the wcte12xp
driver (for the single span VoiceBus-based T1/E1 cards) did not properly
update the 'devicetype' field reported by dahdi_scan when a VPM was found
during the card startup process. As a result, dahdi_scan did not show that
a VPM was present on the card, even if it was.



git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7038 a0bf4364-ded3-4de4-8d8a-66a801d63aff