Remove duplicate-ish code that has existed since r40 (committed 8 years ago).
After a long discussion, it was decided that this was the best action since
there was no apparent issue caused by it.  The alternative suggestion was to
change the extra BBIT to an ABIT as FXSLS uses.  Because of the nature of
this code, it was deemed to be a bit (you see what I did there?) too risky.
Who would've expected such a small change to be so complex?

(closes issue #15178)
Reported by: KrisV


git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7057 a0bf4364-ded3-4de4-8d8a-66a801d63aff