dahdi-base: Do not allow jumps in system time to lock up the system w/core_timer
Since dahdi coretimer uses the number of milliseconds that has actually passed
to determine how many times to call dahdi_receive, it is possible that if the
system time shifts after dahdi is started, that the system can appear to lock
up while the core timer attempts to catch up.  This change prevents soft lock
ups under these conditions.  This is brings the dahdi_dummy changes in r6933
into dahdi-base.

(related to issue #15647)

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7437 a0bf4364-ded3-4de4-8d8a-66a801d63aff