voicebus: Send 'idle' buffers when the transmit descriptor underruns.
Previously, when the host system fails to service the interrupt in a timely
fashion, the transmit descriptor ring for the voicebus card would "go empty"
since the interface wouldn't have another descriptor to read in.  The driver
only knows that it went empty, not how far behind it actually was.  Therefore,
the driver could just increase the latency by a millisecond and keep going
waiting for another bump.

Additionally, when the transmit descriptor actually goes empty, there are some
cases where an in process SPI transaction to one of the modules is interrupted,
which may result in corrupted module register writes on rare occassions.

This now makes it possible for the voicebus drivers to coexist with some devices
that periodically lock interrupts for longer than 25ms.  Before this patch, the
latency would constantly increase until either the modules received a corrupted
frame.

This patch preconfigures all the receive descriptors to send an "idle" packet
that will be transmitted to the onboard modules when the host doesn't service
the interrupt within (latency - 2)ms.  There are now two kinds of underruns,
softunderuns where the driver can detect that these idlebuffers have made it to
the TX FIFO, and the normal hard underrun where the part signals a transmit
descriptor unavailable interrupt.  DAHDI-278.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7517 a0bf4364-ded3-4de4-8d8a-66a801d63aff