DAHDI_AUDIO_NOTIFY: Also pass DAHDI_AUDIOMODE to channel driver
Adds a new compile-time option DAHDI_AUDIO_NOTIFY. With it there's a new
span hook, audio_notify, that gets called (if not NULL) whenever the
ioctl DAHDI_AUDIOMODE is called for a channel.

This can be useful to tell (in most cases) if an ISDN channel is active
or not and thus allow the low-level driver to optimize traffic when
channels are not active.


git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7691 a0bf4364-ded3-4de4-8d8a-66a801d63aff