voicebus: Remove framesize, buffer_cache, and cache_line members from struct.
Framesize is the same on all the clients, so it can be removed from the 'struct
voicebus' (and changed back again if needed).  Since the framesize is the same,
each board doesn't need it's own buffer_cache, and cache_line is only used
during startup and can be read directly from pci configuration space, so we
don't need to keep it around in our context.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@7779 a0bf4364-ded3-4de4-8d8a-66a801d63aff