wctdm24xxp: Only use the first three EFRAMES for module commands.
On a 2.6.32.7 debug kernel with a TDM2400, this can save up to 30ms every second
of processing time without any appreciable impact on runtime performance since
most of the commands to the modules during normal operation are reads, and
therefore only one would be in each SFRAME anyway.  The impact of this change is
less for TDM410 and TDM800 since they already limit the number of modules that
they look for queued commands for.

There is still more room for optimizing module command handling.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8096 a0bf4364-ded3-4de4-8d8a-66a801d63aff