wcte12xp:  Poll the VPM and reset it if necessary.
When the transmit descriptor runs out completely, there (appears to be) a
chance for a random command to be sent that results in the VPMADT032 to no
longer respond, typically resulting in one way audio.  This change introduces
a poll of the VPM.  If it fails the poll, it will be bypassed temporarily
while the driver resets and reprograms it.  Also, the VPM is initially
programmed in the spanconfig callback instead of at driver load.  This moves
the potential for underruns until later in the boot process. DAHDI-573.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8468 a0bf4364-ded3-4de4-8d8a-66a801d63aff