wctc4xxp: Fix "operation on may be undefined" warning.
gcc 4.5.0 was generating a warning on the changed line, and the
discussion at http://gcc.gnu.org/ml/gcc/2004-10/msg00024.html explains
why.

Essentially, the only thing guaranteed with the preincrement operator is
that the value will be incremented before the assignment.  It's
undefined where in the sequence the mask will be applied.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8831 a0bf4364-ded3-4de4-8d8a-66a801d63aff