Don't assume hardwired pri_protocol is E1
When initializing registers for the "PRI" (E1/T1) module, in the case the
user has not explicitly asked for E1/T1 settings in xpp.conf (pri_protocol)
fall back to the current value, which is the build-time fall-back.

The value is read from sysfs.

Previously we implicitly assumed this value is E1.


git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8873 a0bf4364-ded3-4de4-8d8a-66a801d63aff