wcte12xp, wctdm24xxp: Return buffer processing to interrupt handler.
In revision 8095, I had moved most of the buffer processing out of the
interrupt handler and into a tasklet. The intended result was to enable
multiple cards to interleave with one another.  But once again I was
bitten by the fact that there are some systems that for one reason or
another do not process their tasklets in a timely enough manner for the
real-time nature of TDM processing.  This commit moves this processing
back into the interrupt handler by default.  It also limits the number
of frames that the interrupt handler will process at any given time
which appears to achieve the same intended result.

(closes issue #17289)
Tested by: alecdavis

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8982 a0bf4364-ded3-4de4-8d8a-66a801d63aff