dahdi: Remove the 'pvt' member from dahdi_span.
The vast majority of board drivers already keep the dahdi_span structure
in a driver specific structure.  The others were easily converted.  This
way board drivers can use the container_of macro to find what was
previously pointed to by the "pvt" member of the span.  One less thing
to think about in the span structure.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8984 a0bf4364-ded3-4de4-8d8a-66a801d63aff