Added hardware transformer check
QuadFalc has the ability to test if the transformer is
performing correctly. If the components between the framer
and the physical span interface are shorted out or opened
for any reason we can now detect it. Possible causes for
tripping this error could be a broken transformer from
an electrical spike or a board manufacturing error.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@8999 a0bf4364-ded3-4de4-8d8a-66a801d63aff