wctc4xxp: Do not ACK response packets without a corresponding request.
Closes a small window of opportunity where system conditions prevent the
driver from servicing it's receive ring within the timeout period of a
request, and then in the middle of retrying the request (after the
request was already removed from the "waiting_for_response" list) the
driver process the response, it's possible for the driver to ACK the
response without ever pairing it up with the original request. The
result being that the DTE will then ignore our attempts to retry the
original request. DAHDI-430.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@9035 a0bf4364-ded3-4de4-8d8a-66a801d63aff