wctdm24xxp: Prevent FXS Proslic staying in "Forward/Reverse OnHookTransfer" during call [One-liner summary of changes]
Now that the 3 second click after answering the call on an FXS port has been fixed, the side effect is that the Proslic during a call, still has the audio signal paths still powered on.

Reading the Si3215 specs it reads that an extra ~20mA is consumed while in OHT mode. 

(closes issue #17764)
Reported by: alecdavis
Patches: 
      wctdm24xxp_fxs_offhook.diff.txt uploaded by alecdavis (license 585)
Tested by: alecdavis



git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@9070 a0bf4364-ded3-4de4-8d8a-66a801d63aff