dahdi: Better tolerate surprise removal of channels.
Enable DAHDI to detect if an operation on a open file handle refers to a
channel that may have been hot swapped out.

This is done by ensuring that file->private_data is always properly set for
any open channel, and it's set back to NULL when a channel is unregistered.
Then file->private_data can be used to check whether it's valid to perform an
operation on the channel.

(issue #17669)
Reported by: tzafrir
Tested by: sruffell