wctdm24xxp: Added "auto" companding option
Added the "auto" option for the module parameter "companding". In auto
mode its left up to each card to decide what companding mode it should
be in. If a BRI module is installed, it set everything to alaw; ulaw
is selected if no BRI module is installed.

Also fixed a bug where forcing companding one way or the other wasn't
affecting the BRI modules. dahdi-673

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@9101 a0bf4364-ded3-4de4-8d8a-66a801d63aff