wctdm24xxp: Set the companding mode on the analog ports in fixup.
Fixes a regression from 9101 'wctdm24xxp: Added "auto" companding option'
where the analog modules were not defaulted to alaw properly when a
digital module is on an Hx8.  This could result in very poor audio since the
modules were providing ulaw data, but dahdi-base believed the audio was in
alaw when converting to signed linear.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@9143 a0bf4364-ded3-4de4-8d8a-66a801d63aff