wcte12xp, wctdm24xxp: Add compile time option CONFIG_VOICEBUS_ECREFERENCE.
Add compile time option to improve the reference signal provided to
software echo cancelers.  The intent here is for this functionality to
become the default behavior but more testing and work on the edge cases
is needed.  It's being brought in now as a compile time option since there
have been reports that it helps in some environments.

Instead of using two buffers, which means that at best we're two
milliseconds behind, use a circular buffer where audio data is written
on the transmit side and read on the receive path.  In this way high
latency values will not interfere with the operation of software
echo cancelers. DAHDI-291. DAHDI-387.

This work was originally on:
http://svn.asterisk.org/svn/dahdi/linux/team/sruffell/improved_ecreference@9143

and includes a generic kfifo replacement by Matt Fredrickson.

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/trunk@9144 a0bf4364-ded3-4de4-8d8a-66a801d63aff