dahdi: Remove 'chans' array from reg/unreg. Now completely removed.
This change removes the remaining locations the 'chans' array was
referenced, but keeps the observable behavior to the way it was when the
chans array. Namely, channels are numbered in registration order.  Only
now it's possible to renumber channels easily since their number isn't
also their implied location in any arrays.